aws rds connection and database setup
AWS data loading,
load of data increasing and implementing joins,
table schema change

+++ insurance/src/main/java/com/a2z/insurance/model/contact/ContactBankAccounts.java
@@ -18,7 +18,7 @@ import lombok.Setter;
 public class ContactBankAccounts {
     @Id
     private int id;
-    private int contactNumber;
+    private long contactNumber;
     private String bankName;
     private int countryCode;
     private String branchName;
